+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/caching
 
+## 0.9.3
+
+### Patch Changes
+
+- Updated dependencies [32a0720]
+- Updated dependencies [54e117e]
+- Updated dependencies [be6e109]
+  - @cat-factory/kernel@0.134.0
+
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies [6564507]
+  - @cat-factory/kernel@0.133.0
+
 ## 0.9.1
 
 ### Patch Changes