+++ backend/packages/observability-otel/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/observability-otel
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [32a0720]
+- Updated dependencies [54e117e]
+- Updated dependencies [be6e109]
+  - @cat-factory/kernel@0.134.0
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [6564507]
+  - @cat-factory/kernel@0.133.0
+
 ## 0.1.7
 
 ### Patch Changes